refactoring target for getting setup-script

+++ Makefile
@@ -17,13 +17,13 @@ ps:
 logs:
 	@docker-compose logs -f
 
-# https://github.com/tomav/docker-mailserver/wiki/Setup-docker-mailserver-using-the-script-setup.sh
-
-fetch-setup:
-	wget -q -O setup.sh https://raw.githubusercontent.com/tomav/docker-mailserver/master/setup.sh; chmod a+x ./setup.sh
-
 pull:
 	@docker pull tvial/docker-mailserver:${MAILSERVER_TAG}
+
+# https://github.com/tomav/docker-mailserver/wiki/Setup-docker-mailserver-using-the-script-setup.sh
+
+get-setup_script:
+	test -f setup.sh ||  wget -q -O setup.sh https://raw.githubusercontent.com/tomav/docker-mailserver/master/setup.sh; chmod a+x ./setup.sh
 
 setup:
 	./setup.sh  -i tvial/docker-mailserver:${MAILSERVER_TAG} email add akranes@mail.dina-web.net akranes
